pyproject.toml
[tool.poetry]
name = "mch-fastapi"
version = "0.1.0"
description = "i realy havent ideas what the fuck is this"
authors = ["M.Chupakhin <mchupakhinwork@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
argon2-cffi = "^23.1.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
uvicorn = "^0.29.0"
SQLAlchemy = {extras = ["acincio"], version = "^2.0.29"}
websockets= "^12.0"
load-dotenv = "^0.1.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
pydantic-settings = "^2.2.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
